+++ log/docs/changelog.md
@@ -3,6 +3,12 @@ NEXT
 
 - Log: Support `yield from` statements in the auto-logging.
 - Log cli: Properly assign args to generator.
+- Public API change: `critical`/`warn`/`info` methods now accept multiple arguments and all are concatenated and converted to `str`.
+    - Old api: `robocorp.log.info(messsage: str, html: bool=False)`
+    - New api: `robocorp.log.info(*message)`
+- New API to embed html into the page:
+    - `robocorp.log.html(html: str, level: str = "INFO")`
+        - `level` may be `"INFO"`, `"WARN"` or `"ERROR"`
 
 0.0.12
 -----------------------------